Override PSGI configuration by PSGI_FILE, PSGI_URI_PATH variables
This introduces two assmeble-time environment variables to override
PSGI detection:

If PSGI_FILE variable is set but empty, skip it.
If PSGI_FILE is set and non-empty, use it.
If PSGI_FILE does not exist, check if exactly one ./*.psgi file exists and
use that file.

If PSGI_URI_PATH variable has a value, use it as a location. Default is "/".

<https://github.com/openshift/sti-perl/issues/58>