Remove IfModule check
Since we're processing our custom configuration before reading system
config files, the MPM prefork module is not loaded at the time of the
evaluation. Since this image will always run with Prefork (threaded
workers don't work with Perl), we can always set the settings, not
needing to check for the module to be loaded.