Separate --mirror argument from CPAN_MIRROR env var
Currently, users need to add '--mirror' option along with mirror URL
in CPAN_MIRROR env var in order to work correctly. However, that option
'--mirror' shouldn't be a part of CPAN_MIRROR as it's rather confusing
to add an option into an env var that is used for mirror URL only.

Signed-off-by: Vu Dinh <vdinh@redhat.com>